server/requirements.txt
tqdm==4.64.1
uvicorn==0.20.0
SQLAlchemy==2.0.4
python-dotenv==1.0.0
psycopg2-binary
numpy==1.24.2
pandas==1.5.3
fastapi==0.92.0